+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2023-02-14
+
+### Added
+
+* Spring Boot auto configuration module `tw-gaffer-jta-starter`.
+
+### Changed
+
+* Connection and DataSource wrappers are now implementing `tw-base-utils`'s wrapper interfaces.
+
 ## [2.0.0] - 2022-12-13
 
 ### Fixed